Refactor screens so that we can reuse the same find/select table for update/delete operations
